# Conflicts:
#	src/main/java/com/skapp/community/common/constant/CommonMessageConstant.java
#	src/main/java/com/skapp/enterprise
#	src/main/resources/community/messages/common-messages.properties

+++ src/main/java/com/skapp/community/leaveplanner/payload/response/EmployeeLeaveEntitlementReportExportDto.java
@@ -21,10 +21,10 @@ public class EmployeeLeaveEntitlementReportExportDto {
 
 	private String leaveName;
 
-	private Double totalDaysAllocated;
+	private Float totalDaysAllocated;
 
-	private Double totalDaysUsed;
+	private Float totalDaysUsed;
 
-	private Double totalBalanceDays;
+	private Float totalBalanceDays;
 
 }
